+++ src/journal/queries/JournalDateContainerQuery.js
@@ -1,0 +1,12 @@
+import { graphql } from 'react-relay';
+
+const JournalDateContainerQuery = graphql`
+  query JournalDateContainerQuery ($condition: FoodSelectionCondition) {
+    currentPerson {
+      ...JournalDateContainer_currentPerson,
+    }
+  }
+`;
+
+export default JournalDateContainerQuery;
+